refactor: split instr module
