+++ sc62015/pysc62015/instr/__init__.py
@@ -1,0 +1,3 @@
+from .opcodes import *  # noqa: F401,F403
+from .instructions import *  # noqa: F401,F403
+from .opcode_table import OPCODES  # noqa: F401
