+++ sc62015/pysc62015/instr/opcode_table.py
@@ -1,0 +1,288 @@
+from .instructions import *  # noqa: F401,F403
+from .opcodes import *  # noqa: F401,F403
+OPCODES = {
+    # 00h
+    0x00: NOP,
+    0x01: RETI,
+    0x02: (JP_Abs, Opts(ops=[Imm16()])),
+    0x03: (JP_Abs, Opts(name="JPF", ops=[Imm20()])),
+    0x04: (CALL, Opts(ops=[Imm16()])),
+    0x05: (CALL, Opts(name="CALLF", ops=[Imm20()])),
+    0x06: RET,
+    0x07: RETF,
+    0x08: (MV, Opts(ops=[Reg("A"), Imm8()])),
+    0x09: (MV, Opts(ops=[Reg("IL"), Imm8()])),
+    0x0A: (MV, Opts(ops=[Reg("BA"), Imm16()])),
+    0x0B: (MV, Opts(ops=[Reg("I"), Imm16()])),
+    0x0C: (MV, Opts(ops=[Reg("X"), Imm20()])),
+    0x0D: (MV, Opts(ops=[Reg("Y"), Imm20()])),
+    0x0E: (MV, Opts(ops=[Reg("U"), Imm20()])),
+    0x0F: (MV, Opts(ops=[Reg("S"), Imm20()])),
+    # 10h
+    0x10: (JP_Abs, Opts(ops=[IMem20()])),
+    0x11: (JP_Abs, Opts(ops=[Reg3()])),
+    0x12: (JP_Rel, Opts(ops=[ImmOffset("+")])),
+    0x13: (JP_Rel, Opts(ops=[ImmOffset("-")])),
+    0x14: (JP_Abs, Opts(cond="Z", ops=[Imm16()])),
+    0x15: (JP_Abs, Opts(cond="NZ", ops=[Imm16()])),
+    0x16: (JP_Abs, Opts(cond="C", ops=[Imm16()])),
+    0x17: (JP_Abs, Opts(cond="NC", ops=[Imm16()])),
+    0x18: (JP_Rel, Opts(cond="Z", ops=[ImmOffset("+")])),
+    0x19: (JP_Rel, Opts(cond="Z", ops=[ImmOffset("-")])),
+    0x1A: (JP_Rel, Opts(cond="NZ", ops=[ImmOffset("+")])),
+    0x1B: (JP_Rel, Opts(cond="NZ", ops=[ImmOffset("-")])),
+    0x1C: (JP_Rel, Opts(cond="C", ops=[ImmOffset("+")])),
+    0x1D: (JP_Rel, Opts(cond="C", ops=[ImmOffset("-")])),
+    0x1E: (JP_Rel, Opts(cond="NC", ops=[ImmOffset("+")])),
+    0x1F: (JP_Rel, Opts(cond="NC", ops=[ImmOffset("-")])),
+    # 20h
+    0x20: UnknownInstruction,
+    0x21: PRE,
+    0x22: PRE,
+    0x23: PRE,
+    0x24: PRE,
+    0x25: PRE,
+    0x26: PRE,
+    0x27: PRE,
+    0x28: (PUSHU, Opts(ops=[Reg("A")])),
+    0x29: (PUSHU, Opts(ops=[Reg("IL")])),
+    0x2A: (PUSHU, Opts(ops=[Reg("BA")])),
+    0x2B: (PUSHU, Opts(ops=[Reg("I")])),
+    0x2C: (PUSHU, Opts(ops=[Reg("X")])),
+    0x2D: (PUSHU, Opts(ops=[Reg("Y")])),
+    0x2E: (PUSHU, Opts(ops=[RegF()])),
+    0x2F: (PUSHU, Opts(ops=[RegIMR()])),
+    # 30h
+    0x30: PRE,
+    0x31: PRE,
+    0x32: PRE,
+    0x33: PRE,
+    0x34: PRE,
+    0x35: PRE,
+    0x36: PRE,
+    0x37: PRE,
+    0x38: (POPU, Opts(ops=[Reg("A")])),
+    0x39: (POPU, Opts(ops=[Reg("IL")])),
+    0x3A: (POPU, Opts(ops=[Reg("BA")])),
+    0x3B: (POPU, Opts(ops=[Reg("I")])),
+    0x3C: (POPU, Opts(ops=[Reg("X")])),
+    0x3D: (POPU, Opts(ops=[Reg("Y")])),
+    0x3E: (POPU, Opts(ops=[RegF()])),
+    0x3F: (POPU, Opts(ops=[RegIMR()])),
+    # 40h
+    0x40: (ADD, Opts(ops=[Reg("A"), Imm8()])),
+    0x41: (ADD, Opts(ops=[IMem8(), Imm8()])),
+    0x42: (ADD, Opts(ops=[Reg("A"), IMem8()])),
+    0x43: (ADD, Opts(ops=[IMem8(), Reg("A")])),
+    0x44: (ADD, Opts(ops=[RegPair(size=2)])),
+    0x45: (ADD, Opts(ops=[RegPair(size=3)])),
+    0x46: (ADD, Opts(ops=[RegPair(size=1)])),
+    0x47: (PMDF, Opts(ops=[IMem8(), Imm8()])),
+    0x48: (SUB, Opts(ops=[Reg("A"), Imm8()])),
+    0x49: (SUB, Opts(ops=[IMem8(), Imm8()])),
+    0x4A: (SUB, Opts(ops=[Reg("A"), IMem8()])),
+    0x4B: (SUB, Opts(ops=[IMem8(), Reg("A")])),
+    0x4C: (SUB, Opts(ops=[RegPair(size=2)])),
+    0x4D: (SUB, Opts(ops=[RegPair(size=3)])),
+    0x4E: (SUB, Opts(ops=[RegPair(size=1)])),
+    0x4F: (PUSHS, Opts(ops=[RegF()])),
+    # 50h
+    0x50: (ADC, Opts(ops=[Reg("A"), Imm8()])),
+    0x51: (ADC, Opts(ops=[IMem8(), Imm8()])),
+    0x52: (ADC, Opts(ops=[Reg("A"), IMem8()])),
+    0x53: (ADC, Opts(ops=[IMem8(), Reg("A")])),
+    0x54: (ADCL, Opts(ops=[IMem8(), IMem8()])),
+    0x55: (ADCL, Opts(ops=[IMem8(), Reg("A")])),
+    0x56: (MVL, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM,
+                                        allowed_modes=[EMemRegMode.POSITIVE_OFFSET, EMemRegMode.NEGATIVE_OFFSET])])),
+    0x57: (PMDF, Opts(ops=[IMem8(), Reg("A")])),
+    0x58: (SBC, Opts(ops=[Reg("A"), Imm8()])),
+    0x59: (SBC, Opts(ops=[IMem8(), Imm8()])),
+    0x5A: (SBC, Opts(ops=[Reg("A"), IMem8()])),
+    0x5B: (SBC, Opts(ops=[IMem8(), Reg("A")])),
+    0x5C: (SBCL, Opts(ops=[IMem8(), IMem8()])),
+    0x5D: (SBCL, Opts(ops=[IMem8(), Reg("A")])),
+    0x5E: (MVL, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET,
+                                        allowed_modes=[EMemRegMode.POSITIVE_OFFSET, EMemRegMode.NEGATIVE_OFFSET])])),
+    0x5F: (POPS, Opts(ops=[RegF()])),
+    # 60h
+    0x60: (CMP, Opts(ops=[Reg("A"), Imm8()])),
+    0x61: (CMP, Opts(ops=[IMem8(), Imm8()])),
+    0x62: (CMP, Opts(ops=[EMemAddr(width=1), Imm8()])),
+    0x63: (CMP, Opts(ops=[IMem8(), Reg("A")])),
+    0x64: (TEST, Opts(ops=[Reg("A"), Imm8()])),
+    0x65: (TEST, Opts(ops=[IMem8(), Imm8()])),
+    0x66: (TEST, Opts(ops=[EMemAddr(width=1), Imm8()])),
+    0x67: (TEST, Opts(ops=[IMem8(), Reg("A")])),
+    0x68: (XOR, Opts(ops=[Reg("A"), Imm8()])),
+    0x69: (XOR, Opts(ops=[IMem8(), Imm8()])),
+    0x6A: (XOR, Opts(ops=[EMemAddr(width=1), Imm8()])),
+    0x6B: (XOR, Opts(ops=[IMem8(), Reg("A")])),
+    0x6C: (INC, Opts(ops=[Reg3()])),
+    0x6D: (INC, Opts(ops=[IMem8()])),
+    0x6E: (XOR, Opts(ops=[IMem8(), IMem8()])),
+    0x6F: (XOR, Opts(ops=[Reg("A"), IMem8()])),
+    # 70h
+    0x70: (AND, Opts(ops=[Reg("A"), Imm8()])),
+    0x71: (AND, Opts(ops=[IMem8(), Imm8()])),
+    0x72: (AND, Opts(ops=[EMemAddr(width=1), Imm8()])),
+    0x73: (AND, Opts(ops=[IMem8(), Reg("A")])),
+    0x74: (MV, Opts(ops=[Reg("A"), RegB()])),
+    0x75: (MV, Opts(ops=[RegB(), Reg("A")])),
+    0x76: (AND, Opts(ops=[IMem8(), IMem8()])),
+    0x77: (AND, Opts(ops=[Reg("A"), IMem8()])),
+    0x78: (OR, Opts(ops=[Reg("A"), Imm8()])),
+    0x79: (OR, Opts(ops=[IMem8(), Imm8()])),
+    0x7A: (OR, Opts(ops=[EMemAddr(width=1), Imm8()])),
+    0x7B: (OR, Opts(ops=[IMem8(), Reg("A")])),
+    0x7C: (DEC, Opts(ops=[Reg3()])),
+    0x7D: (DEC, Opts(ops=[IMem8()])),
+    0x7E: (OR, Opts(ops=[IMem8(), IMem8()])),
+    0x7F: (OR, Opts(ops=[Reg("A"), IMem8()])),
+    # 80h
+    0x80: (MV, Opts(ops=[Reg("A"), IMem8()])),
+    0x81: (MV, Opts(ops=[Reg("IL"), IMem8()])),
+    0x82: (MV, Opts(ops=[Reg("BA"), IMem16()])),
+    0x83: (MV, Opts(ops=[Reg("I"), IMem16()])),
+    0x84: (MV, Opts(ops=[Reg("X"), IMem20()])),
+    0x85: (MV, Opts(ops=[Reg("Y"), IMem20()])),
+    0x86: (MV, Opts(ops=[Reg("U"), IMem20()])),
+    0x87: (MV, Opts(ops=[Reg("S"), IMem20()])),
+    0x88: (MV, Opts(ops=[Reg("A"), EMemAddr(width=1)])),
+    0x89: (MV, Opts(ops=[Reg("IL"), EMemAddr(width=1)])),
+    0x8A: (MV, Opts(ops=[Reg("BA"), EMemAddr(width=2)])),
+    0x8B: (MV, Opts(ops=[Reg("I"), EMemAddr(width=2)])),
+    0x8C: (MV, Opts(ops=[Reg("X"), EMemAddr(width=3)])),
+    0x8D: (MV, Opts(ops=[Reg("Y"), EMemAddr(width=3)])),
+    0x8E: (MV, Opts(ops=[Reg("U"), EMemAddr(width=3)])),
+    0x8F: (MV, Opts(ops=[Reg("S"), EMemAddr(width=3)])),
+    # 90h
+    0x90: (MV, Opts(ops=[Reg("A"), EMemReg(width=1)])),
+    0x91: (MV, Opts(ops=[Reg("IL"), EMemReg(width=1)])),
+    0x92: (MV, Opts(ops=[Reg("BA"), EMemReg(width=2)])),
+    0x93: (MV, Opts(ops=[Reg("I"), EMemReg(width=2)])),
+    0x94: (MV, Opts(ops=[Reg("X"), EMemReg(width=3)])),
+    0x95: (MV, Opts(ops=[Reg("Y"), EMemReg(width=3)])),
+    0x96: (MV, Opts(ops=[Reg("U"), EMemReg(width=3)])),
+    0x97: SC,
+    0x98: (MV, Opts(ops=[Reg("A"), EMemIMem()])),
+    0x99: (MV, Opts(ops=[Reg("IL"), EMemIMem()])),
+    0x9A: (MV, Opts(ops=[Reg("BA"), EMemIMem()])),
+    0x9B: (MV, Opts(ops=[Reg("I"), EMemIMem()])),
+    0x9C: (MV, Opts(ops=[Reg("X"), EMemIMem()])),
+    0x9D: (MV, Opts(ops=[Reg("Y"), EMemIMem()])),
+    0x9E: (MV, Opts(ops=[Reg("U"), EMemIMem()])),
+    0x9F: RC,
+    # A0h
+    0xA0: (MV, Opts(ops=[IMem8(), Reg("A")])),
+    0xA1: (MV, Opts(ops=[IMem8(), Reg("IL")])),
+    0xA2: (MV, Opts(ops=[IMem16(), Reg("BA")])),
+    0xA3: (MV, Opts(ops=[IMem16(), Reg("I")])),
+    0xA4: (MV, Opts(ops=[IMem20(), Reg("X")])),
+    0xA5: (MV, Opts(ops=[IMem20(), Reg("Y")])),
+    0xA6: (MV, Opts(ops=[IMem20(), Reg("U")])),
+    0xA7: (MV, Opts(ops=[IMem20(), Reg("S")])),
+    0xA8: (MV, Opts(ops=[EMemAddr(width=1), Reg("A")])),
+    0xA9: (MV, Opts(ops=[EMemAddr(width=1), Reg("IL")])),
+    0xAA: (MV, Opts(ops=[EMemAddr(width=2), Reg("BA")])),
+    0xAB: (MV, Opts(ops=[EMemAddr(width=2), Reg("I")])),
+    0xAC: (MV, Opts(ops=[EMemAddr(width=3), Reg("X")])),
+    0xAD: (MV, Opts(ops=[EMemAddr(width=3), Reg("Y")])),
+    0xAE: (MV, Opts(ops=[EMemAddr(width=3), Reg("U")])),
+    0xAF: (MV, Opts(ops=[EMemAddr(width=3), Reg("S")])),
+    # B0h
+    0xB0: (MV, Opts(ops=[EMemReg(width=1), Reg("A")])),
+    0xB1: (MV, Opts(ops=[EMemReg(width=1), Reg("IL")])),
+    0xB2: (MV, Opts(ops=[EMemReg(width=2), Reg("BA")])),
+    0xB3: (MV, Opts(ops=[EMemReg(width=2), Reg("I")])),
+    0xB4: (MV, Opts(ops=[EMemReg(width=3), Reg("X")])),
+    0xB5: (MV, Opts(ops=[EMemReg(width=3), Reg("Y")])),
+    0xB6: (MV, Opts(ops=[EMemReg(width=3), Reg("U")])),
+    0xB7: (CMP, Opts(ops=[IMem8(), IMem8()])),
+    0xB8: (MV, Opts(ops=[EMemIMem(), Reg("A")])),
+    0xB9: (MV, Opts(ops=[EMemIMem(), Reg("IL")])),
+    0xBA: (MV, Opts(ops=[EMemIMem(), Reg("BA")])),
+    0xBB: (MV, Opts(ops=[EMemIMem(), Reg("I")])),
+    0xBC: (MV, Opts(ops=[EMemIMem(), Reg("X")])),
+    0xBD: (MV, Opts(ops=[EMemIMem(), Reg("Y")])),
+    0xBE: (MV, Opts(ops=[EMemIMem(), Reg("U")])),
+    0xBF: UnknownInstruction,
+    # C0h
+    0xC0: (EX, Opts(ops=[IMem8(), IMem8()])),
+    0xC1: (EX, Opts(name="EXW", ops=[IMem16(), IMem16()])),
+    0xC2: (EX, Opts(name="EXP", ops=[IMem20(), IMem20()])),
+    0xC3: (EXL, Opts(ops=[IMem8(), IMem8()])),
+    0xC4: (DADL, Opts(ops=[IMem8(), IMem8()])),
+    0xC5: (DADL, Opts(ops=[IMem8(), Reg("A")])),
+    0xC6: (CMPW, Opts(ops=[IMem16(), IMem16()])),
+    0xC7: (CMPP, Opts(ops=[IMem20(), IMem20()])),
+    0xC8: (MV, Opts(ops=[IMem8(), IMem8()])),
+    0xC9: (MV, Opts(name="MVW", ops=[IMem16(), IMem16()])),
+    0xCA: (MV, Opts(name="MVP", ops=[IMem20(), IMem20()])),
+    0xCB: (MVL, Opts(ops=[IMem8(), IMem8()])),
+    0xCC: (MV, Opts(ops=[IMem8(), Imm8()])),
+    0xCD: (MV, Opts(name="MVW", ops=[IMem16(), Imm16()])),
+    0xCE: TCL,
+    0xCF: (MVLD, Opts(ops=[IMem8(), IMem8()])),
+    # D0h
+    0xD0: (MV, Opts(ops=[IMem8(), EMemAddr(width=1)])),
+    0xD1: (MV, Opts(name="MVW", ops=[IMem16(), EMemAddr(width=2)])),
+    0xD2: (MV, Opts(name="MVP", ops=[IMem20(), EMemAddr(width=3)])),
+    0xD3: (MVL, Opts(ops=[IMem8(), EMemAddr(width=1)])),
+    0xD4: (DSBL, Opts(ops=[IMem8(), IMem8()])),
+    0xD5: (DSBL, Opts(ops=[IMem8(), Reg("A")])),
+    0xD6: (CMPW, Opts(ops_reversed=True, ops=[IMem16(), Reg3()])),
+    0xD7: (CMPP, Opts(ops_reversed=True, ops=[IMem20(), Reg3()])),
+    0xD8: (MV, Opts(ops=[EMemAddr(width=1), IMem8()])),
+    0xD9: (MV, Opts(name="MVW", ops=[EMemAddr(width=2), IMem16()])),
+    0xDA: (MV, Opts(name="MVP", ops=[EMemAddr(width=3), IMem20()])),
+    0xDB: (MVL, Opts(ops=[EMemAddr(width=1), IMem8()])),
+    0xDC: (MV, Opts(name="MVP", ops=[IMem20(), Imm20()])),
+    0xDD: (EX, Opts(ops=[Reg("A"), RegB()])),
+    0xDE: HALT,
+    0xDF: OFF,
+    # E0h
+    0xE0: (MV, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
+    0xE1: (MV, Opts(name="MVW",
+                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
+    0xE2: (MV, Opts(name="MVP",
+                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
+    # FIXME: verify width
+    0xE3: (MVL, Opts(ops_reversed=True, ops=[IMem8(),
+                                             EMemReg(width=1,
+                                                     allowed_modes=[EMemRegMode.POST_INC,
+                                                                    EMemRegMode.PRE_DEC])])),
+    0xE4: (ROR, Opts(ops=[Reg("A")])),
+    0xE5: (ROR, Opts(ops=[IMem8()])),
+    0xE6: (ROL, Opts(ops=[Reg("A")])),
+    0xE7: (ROL, Opts(ops=[IMem8()])),
+    0xE8: (MV, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
+    0xE9: (MV, Opts(name="MVW",
+                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
+    0xEA: (MV, Opts(name="MVP",
+                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
+    # FIXME: verify width
+    0xEB: (MVL, Opts(ops=[EMemReg(width=1), IMem8()])),
+    0xEC: (DSLL, Opts(ops=[IMem8()])),
+    0xED: (EX, Opts(ops=[RegPair(size=2)])),
+    0xEE: (SWAP, Opts(ops=[Reg("A")])),
+    0xEF: WAIT,
+    # F0h
+    0xF0: (MV, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
+    0xF1: (MV, Opts(name='MVW', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
+    0xF2: (MV, Opts(name='MVP', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
+    0xF3: (MVL, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
+    0xF4: (SHR, Opts(ops=[Reg("A")])),
+    0xF5: (SHR, Opts(ops=[IMem8()])),
+    0xF6: (SHL, Opts(ops=[Reg("A")])),
+    0xF7: (SHL, Opts(ops=[IMem8()])),
+    0xF8: (MV, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
+    0xF9: (MV, Opts(name='MVW', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
+    0xFA: (MV, Opts(name='MVP', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
+    0xFB: (MVL, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
+    0xFC: (DSRL, Opts(ops=[IMem8()])),
+    0xFD: (MV, Opts(ops=[RegPair(size=2)])),
+    0xFE: IR,
+    0xFF: RESET,
+}
+
