+++ sc62015/pysc62015/instr/opcodes.py
@@ -1,18 +1,18 @@
 # based on https://github.com/whitequark/binja-avnera/blob/main/mc/instr.py
-from .tokens import Token, TInstr, TText, TSep, TInt, TReg, TBegMem, TEndMem, MemType
-from .coding import Decoder, Encoder, BufferTooShort
-from .mock_analysis import BranchType
-from .mock_llil import MockLLIL
-from .constants import INTERNAL_MEMORY_START
+from ..tokens import Token, TInstr, TText, TSep, TInt, TReg, TBegMem, TEndMem, MemType
+from ..coding import Decoder, Encoder, BufferTooShort
+
+from ..mock_llil import MockLLIL
+from ..constants import INTERNAL_MEMORY_START
 
 import copy
 from dataclasses import dataclass
-from typing import Optional, List, Generator, Iterator, Dict, Tuple, Union, Type, Literal, Any, Callable, cast
+from typing import Optional, List, Generator, Iterator, Dict, Tuple, Union, Type, Literal, Any, cast
 import enum
 from contextlib import contextmanager
 
 
-from . import binja_api # noqa: F401
+from .. import binja_api # noqa: F401
 from binaryninja import (
     InstructionInfo,
 )
@@ -1621,1142 +1621,4 @@ def lift_loop(il: LowLevelILFunction) -> Generator[None, None, None]:
     il.append(il.if_expr(cond, if_true, if_false))
     il.mark_label(if_true)
 
-
-class NOP(Instruction):
-     def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.nop())
-
-class JumpInstruction(Instruction):
-    def lift_jump_addr(self, il: LowLevelILFunction, addr: int) -> ExpressionIndex:
-        raise NotImplementedError("lift_jump_addr() not implemented")
-
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-        # expect TrueBranch to be handled by subclasses as it might require
-        # llil logic to calculate the address
-        info.add_branch(BranchType.FalseBranch, addr + self.length())
-
-
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        if_true  = LowLevelILLabel()
-        if_false = LowLevelILLabel()
-
-        if self._cond:
-            zero = il.const(1, 0)
-            one  = il.const(1, 1)
-            flag = il.flag(ZFlag) if "Z" in self._cond else il.flag(CFlag)
-            value = zero if "N" in self._cond else one
-
-            cond = il.compare_equal(1, flag, value)
-            il.append(il.if_expr(cond, if_true, if_false))
-
-        il.mark_label(if_true)
-        il.append(il.jump(self.lift_jump_addr(il, addr)))
-        il.mark_label(if_false)
-
-
-class JP_Abs(JumpInstruction):
-    def name(self) -> str:
-        return super().name() + (self._cond if self._cond else "")
-
-    def lift_jump_addr(self, il: LowLevelILFunction, addr: int) -> ExpressionIndex:
-        first, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        assert isinstance(first, HasWidth), f"Expected HasWidth, got {type(first)}"
-        if first.width() >= 3:
-            return first.lift(il)
-        high_addr = addr & 0xFF0000
-        return il.or_expr(3, first.lift(il), il.const(3, high_addr))
-
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-
-        first, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        if isinstance(first, ImmOperand):
-            # absolute address
-            assert first.value is not None, "Value not set"
-            dest = first.value
-            info.add_branch(BranchType.TrueBranch, dest)
-
-class JP_Rel(JumpInstruction):
-    def name(self) -> str:
-        return "JR" + (self._cond if self._cond else "")
-
-    def lift_jump_addr(self, il: LowLevelILFunction, addr: int) -> ExpressionIndex:
-        first, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        assert isinstance(first, ImmOffset), f"Expected ImmOffset, got {type(first)}"
-        return il.const(3, addr + self.length() + first.offset_value())
-
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-        first, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        assert isinstance(first, ImmOffset), f"Expected ImmOffset, got {type(first)}"
-        dest = addr + self.length() + first.offset_value()
-        info.add_branch(BranchType.TrueBranch, dest)
-
-class CALL(Instruction):
-    def _dest(self) -> ImmOperand:
-        dest, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        assert isinstance(dest, ImmOperand), "Expected ImmOperand"
-        return dest
-
-    def dest_addr(self, addr: int) -> int:
-        dest = self._dest()
-        result = dest.value
-        assert result is not None, "Value not set"
-        if dest.width() != 3:
-            assert dest.width() == 2
-            result = addr & 0xFF0000 | result
-        return result
-
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-        info.add_branch(BranchType.CallDestination, self.dest_addr(addr))
-
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dest = self._dest()
-        if dest.width() == 3:
-            il.append(il.call(il.const_pointer(3, self.dest_addr(addr))))
-            return
-        # manually push 2 bytes of address + self.length()
-        il.append(il.push(2, il.const(2, addr + self.length())))
-        il.append(il.jump(il.const_pointer(3, self.dest_addr(addr))))
-
-
-class RetInstruction(Instruction):
-    def addr_size(self) -> int:
-        return 2
-
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-        info.add_branch(BranchType.FunctionReturn)
-
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        # FIXME: should add bitmask for 2-byte pop?
-        il.append(il.ret(il.pop(self.addr_size())))
-
-class RET(RetInstruction): pass
-class RETF(RetInstruction):
-    def addr_size(self) -> int:
-        return 3
-class RETI(RetInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        imr, *rest = RegIMR().operands()
-        imr.lift_assign(il, il.pop(1))
-        RegF().lift_assign(il, il.pop(1))
-        il.append(il.ret(il.pop(3)))
-
-
-class MoveInstruction(Instruction):
-    pass
-
-class MV(MoveInstruction):
-    pass
-
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il_arg2
-
-class MVL(MoveInstruction):
-    def modify_addr_il(self, il: LowLevelILFunction) -> Callable[[int, ExpressionIndex, ExpressionIndex], ExpressionIndex]:
-        return il.add
-
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst, src = self.operands()
-        assert isinstance(dst, Pointer), f"Expected Pointer, got {type(dst)}"
-        assert isinstance(src, Pointer), f"Expected Pointer, got {type(src)}"
-        # 0xCB and 0xCF variants use IMem8, IMem8
-        dst_reg = TempReg(TempMvlDst)
-        dst_mode = get_addressing_mode(self._pre, 1) if self._pre else None
-        src_mode = get_addressing_mode(self._pre, 2) if self._pre else None
-
-        dst_reg.lift_assign(
-            il, dst.lift_current_addr(il, pre=dst_mode, side_effects=False)
-        )
-        src_reg = TempReg(TempMvlSrc)
-        src_reg.lift_assign(
-            il, src.lift_current_addr(il, pre=src_mode, side_effects=False)
-        )
-
-        with lift_loop(il):
-            src_mem = src.memory_helper()(1, src_reg)
-            dst_mem = dst.memory_helper()(1, dst_reg)
-            dst_mem.lift_assign(il, src_mem.lift(il))
-
-            # +1 index
-            func = self.modify_addr_il(il)
-            dst_reg.lift_assign(il, func(dst_reg.width(), dst_reg.lift(il), il.const(1, 1)))
-            src_reg.lift_assign(il, func(src_reg.width(), src_reg.lift(il), il.const(1, 1)))
-
-            # in case we have POST_INC or PRE_DEC, we need to update the
-            # register by lifting it and not assigning it
-            dst.lift_current_addr(il, pre=dst_mode)
-            src.lift_current_addr(il, pre=src_mode)
-
-class MVLD(MVL):
-    def modify_addr_il(self, il: LowLevelILFunction) -> Callable[[int, ExpressionIndex, ExpressionIndex], ExpressionIndex]:
-        return il.sub
-
-class PRE(Instruction):
-    def name(self) -> str:
-        return f"PRE{self.opcode:02x}"
-
-    def fuse(self, sister: 'Instruction') -> Optional['Instruction']:
-        if isinstance(sister, PRE):
-            return None
-        sister._pre = self.opcode
-        sister.set_length(self.length() + sister.length())
-        return sister
-
-class StackInstruction(Instruction):
-    def reg(self) -> Operand:
-        r, *rest = self.operands()
-        assert len(rest) == 0, "Expected no extra operands"
-        return r
-class StackPushInstruction(StackInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        r = self.reg()
-        assert isinstance(r, HasWidth), f"Expected HasWidth, got {type(r)}"
-        il.append(il.push(r.width(), r.lift(il)))
-class StackPopInstruction(StackInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        r = self.reg()
-        assert isinstance(r, HasWidth), f"Expected HasWidth, got {type(r)}"
-        r.lift_assign(il, il.pop(r.width()))
-
-# FIXME: should use U pointer, not S
-class PUSHU(StackInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        r = self.reg()
-        assert isinstance(r, HasWidth)
-        size = r.width()
-        # save the original U so the store uses the pre-decremented value
-        old_u = TempReg(TempIncDecHelper, width=3)
-        old_u.lift_assign(il, il.reg(3, RegisterName("U")))
-        new_u = il.sub(3, old_u.lift(il), il.const(3, size))
-        il.append(il.set_reg(3, RegisterName("U"), new_u))
-        il.append(il.store(size, new_u, r.lift(il)))
-        if isinstance(r, RegIMR):
-            r.lift_assign(il, il.and_expr(1, r.lift(il), il.const(1, 0x7F)))
-
-class POPU(StackInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        r = self.reg()
-        assert isinstance(r, HasWidth)
-        size = r.width()
-        # preserve the pointer prior to increment so the load happens at
-        # the original U value
-        old_u = TempReg(TempIncDecHelper, width=3)
-        old_u.lift_assign(il, il.reg(3, RegisterName("U")))
-        r.lift_assign(il, il.load(size, old_u.lift(il)))
-        il.append(
-            il.set_reg(3, RegisterName("U"), il.add(3, old_u.lift(il), il.const(3, size)))
-        )
-
-class PUSHS(StackPushInstruction): pass
-class POPS(StackPopInstruction): pass
-
-class ArithmeticInstruction(Instruction):
-    def width(self) -> int:
-        first, second = self.operands()
-        assert isinstance(first, HasWidth), f"Expected HasWidth, got {type(first)}"
-        return first.width()
-class ADD(ArithmeticInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.add(self.width(), il_arg1, il_arg2, CZFlag)
-class ADC(ArithmeticInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.add(self.width(), il_arg1, il.add(self.width(), il_arg2,
-                                                    il.flag(CFlag)), CZFlag)
-class SUB(ArithmeticInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.sub(self.width(), il_arg1, il_arg2, CZFlag)
-class SBC(ArithmeticInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.sub(self.width(), il_arg1, il.add(self.width(), il_arg2,
-                                                    il.flag(CFlag)), CZFlag)
-
-
-def bcd_add_emul(il: LowLevelILFunction, w: int, a: ExpressionIndex, b: ExpressionIndex) -> Operand:
-    assert w == 1, "BCD add currently only supports 1-byte operands"
-
-    # Incoming CFlag is the BCD carry from the previous byte's BCD addition
-    incoming_carry = il.flag(CFlag)
-
-    # Low nibble addition: (a & 0xF) + (b & 0xF) + incoming_carry_byte
-    a_low = il.and_expr(1, a, il.const(1, 0x0F))
-    b_low = il.and_expr(1, b, il.const(1, 0x0F))
-    sum_low_nibbles_val = il.add(1, a_low, b_low)
-    sum_low_with_carry_val = il.add(1, sum_low_nibbles_val, incoming_carry) # Max val 9+9+1 = 19 (0x13)
-
-    # Adjust if low nibble sum > 9
-    temp_sum_low_final_reg = TempReg(TempBcdLowNibbleProcessing, width=1)
-    adj_low_needed = il.compare_unsigned_greater_than(1, sum_low_with_carry_val, il.const(1, 9))
-    sum_low_adjusted_val = il.add(1, sum_low_with_carry_val, il.const(1, 0x06))
-
-    label_adj_low_true = LowLevelILLabel()
-    label_adj_low_false = LowLevelILLabel()
-    label_after_adj_low = LowLevelILLabel()
-    il.append(il.if_expr(adj_low_needed, label_adj_low_true, label_adj_low_false))
-    il.mark_label(label_adj_low_true)
-    temp_sum_low_final_reg.lift_assign(il, sum_low_adjusted_val)
-    il.append(il.goto(label_after_adj_low))
-    il.mark_label(label_adj_low_false)
-    temp_sum_low_final_reg.lift_assign(il, sum_low_with_carry_val)
-    il.mark_label(label_after_adj_low)
-
-    current_sum_low_final = temp_sum_low_final_reg.lift(il)
-    result_low_nibble_val = il.and_expr(1, current_sum_low_final, il.const(1, 0x0F))
-    carry_to_high_nibble_val = il.logical_shift_right(1, current_sum_low_final, il.const(1, 4)) # 0 or 1
-
-    # High nibble addition: (a >> 4) + (b >> 4) + carry_to_high_nibble_val
-    a_high = il.logical_shift_right(1, a, il.const(1, 4))
-    b_high = il.logical_shift_right(1, b, il.const(1, 4))
-    sum_high_nibbles_val = il.add(1, a_high, b_high)
-    sum_high_with_carry_val = il.add(1, sum_high_nibbles_val, carry_to_high_nibble_val) # Max 9+9+1 = 19 (0x13)
-
-    # Adjust if high nibble sum > 9
-    temp_sum_high_final_reg = TempReg(TempBcdHighNibbleProcessing, width=1)
-    adj_high_needed = il.compare_unsigned_greater_than(1, sum_high_with_carry_val, il.const(1, 9))
-    sum_high_adjusted_val = il.add(1, sum_high_with_carry_val, il.const(1, 0x06))
-
-    label_adj_high_true = LowLevelILLabel()
-    label_adj_high_false = LowLevelILLabel()
-    label_after_adj_high = LowLevelILLabel()
-    il.append(il.if_expr(adj_high_needed, label_adj_high_true, label_adj_high_false))
-    il.mark_label(label_adj_high_true)
-    temp_sum_high_final_reg.lift_assign(il, sum_high_adjusted_val)
-    il.append(il.goto(label_after_adj_high))
-    il.mark_label(label_adj_high_false)
-    temp_sum_high_final_reg.lift_assign(il, sum_high_with_carry_val)
-    il.mark_label(label_after_adj_high)
-
-    current_sum_high_final = temp_sum_high_final_reg.lift(il)
-    result_high_nibble_val = il.and_expr(1, current_sum_high_final, il.const(1, 0x0F))
-    new_bcd_carry_out_byte_val = il.logical_shift_right(1, current_sum_high_final, il.const(1, 4)) # 0 or 1
-
-    result_byte_val = il.or_expr(1, il.shift_left(1, result_high_nibble_val, il.const(1, 4)), result_low_nibble_val)
-
-    output_reg = TempReg(TempBcdAddEmul, width=1)
-    output_reg.lift_assign(il, result_byte_val)
-    il.append(il.set_flag(CFlag, new_bcd_carry_out_byte_val))
-    # Z flag for current byte (overall Z handled by lift_multi_byte)
-    il.append(il.set_flag(ZFlag, il.compare_equal(1, result_byte_val, il.const(1,0))))
-
-    return output_reg
-
-def bcd_sub_emul(il: LowLevelILFunction, w: int, a: ExpressionIndex, b: ExpressionIndex) -> Operand:
-    assert w == 1, "BCD sub currently only supports 1-byte operands"
-
-    incoming_borrow = il.flag(CFlag) # 0 for no borrow, 1 for borrow
-
-    # Low nibble subtraction: (a_low) - (b_low) - incoming_borrow
-    a_low = il.and_expr(1, a, il.const(1, 0x0F))
-    b_low = il.and_expr(1, b, il.const(1, 0x0F))
-
-    sub_val_low = il.add(1, b_low, incoming_borrow) # bL + Cin
-    temp_sub_low_val = il.sub(1, a_low, sub_val_low)
-
-    # Check for borrow from low nibble
-    borrow_from_low_val = il.compare_signed_less_than(1, temp_sub_low_val, il.const(1, 0))
-
-    final_low_nibble_reg = TempReg(TempBcdLowNibbleProcessing, width=1)
-    adj_val_low = il.sub(1, temp_sub_low_val, il.const(1, 0x06)) # Subtract 6 if borrow
-
-    label_adj_low_true_s = LowLevelILLabel()
-    label_adj_low_false_s = LowLevelILLabel()
-    label_after_adj_low_s = LowLevelILLabel()
-    il.append(il.if_expr(borrow_from_low_val, label_adj_low_true_s, label_adj_low_false_s))
-    il.mark_label(label_adj_low_true_s)
-    final_low_nibble_reg.lift_assign(il, adj_val_low)
-    il.append(il.goto(label_after_adj_low_s))
-    il.mark_label(label_adj_low_false_s)
-    final_low_nibble_reg.lift_assign(il, temp_sub_low_val)
-    il.mark_label(label_after_adj_low_s)
-
-    result_low_nibble_val = il.and_expr(1, final_low_nibble_reg.lift(il), il.const(1, 0x0F))
-
-    # High nibble subtraction: (a_high) - (b_high) - borrow_from_low_val
-    a_high = il.logical_shift_right(1, a, il.const(1, 4))
-    b_high = il.logical_shift_right(1, b, il.const(1, 4))
-
-    sub_val_high = il.add(1, b_high, borrow_from_low_val) # bH + borrow_low
-    temp_sub_high_val = il.sub(1, a_high, sub_val_high)
-
-    new_bcd_borrow_out_byte_val = il.compare_signed_less_than(1, temp_sub_high_val, il.const(1, 0))
-    final_high_nibble_reg = TempReg(TempBcdHighNibbleProcessing, width=1)
-    adj_val_high = il.sub(1, temp_sub_high_val, il.const(1, 0x06))
-
-    label_adj_high_true_s = LowLevelILLabel()
-    label_adj_high_false_s = LowLevelILLabel()
-    label_after_adj_high_s = LowLevelILLabel()
-    il.append(il.if_expr(new_bcd_borrow_out_byte_val, label_adj_high_true_s, label_adj_high_false_s))
-    il.mark_label(label_adj_high_true_s)
-    final_high_nibble_reg.lift_assign(il, adj_val_high)
-    il.append(il.goto(label_after_adj_high_s))
-    il.mark_label(label_adj_high_false_s)
-    final_high_nibble_reg.lift_assign(il, temp_sub_high_val)
-    il.mark_label(label_after_adj_high_s)
-
-    result_high_nibble_val = il.and_expr(1, final_high_nibble_reg.lift(il), il.const(1, 0x0F))
-    result_byte_val = il.or_expr(1, il.shift_left(1, result_high_nibble_val, il.const(1, 4)), result_low_nibble_val)
-
-    output_reg = TempReg(TempBcdSubEmul, width=1)
-    output_reg.lift_assign(il, result_byte_val)
-    il.append(il.set_flag(CFlag, new_bcd_borrow_out_byte_val)) # C=1 if borrow
-    il.append(il.set_flag(ZFlag, il.compare_equal(1, result_byte_val, il.const(1,0))))
-
-    return output_reg
-
-
-def lift_multi_byte(
-    il: LowLevelILFunction,
-    op1: Operand,
-    op2: Operand,
-    clear_carry: bool = False,
-    reverse: bool = False,
-    bcd: bool = False,
-    subtract: bool = False,
-    pre: Optional[int] = None,
-) -> None:
-    assert isinstance(op1, HasWidth), f"Expected HasWidth, got {type(op1)}"
-
-    dst_mode = get_addressing_mode(pre, 1) if pre else None
-    src_mode = get_addressing_mode(pre, 2) if pre else None
-
-    # Helper to create load/store/advance logic for operands
-    def make_handlers(
-        op: Operand,
-        is_dest_op: bool,
-        mode: Optional[AddressingMode],
-    ) -> Tuple[Callable[[], ExpressionIndex],
-                     Callable[[ExpressionIndex], None],
-                     Callable[[], None]]:
-        if isinstance(op, Pointer):
-            # Temp reg to hold the iterating pointer for memory operands
-            ptr_temp_reg_const = TempMultiByte1 if is_dest_op else TempMultiByte2
-            ptr = TempReg(ptr_temp_reg_const, width=3) # Addresses are 3 bytes (20/24 bit)
-
-            # Initialize the pointer temp reg with the initial address from the operand
-            # side_effects=False for source, potentially True for dest if pre/post inc/dec
-            ptr.lift_assign(
-                il,
-                op.lift_current_addr(il, pre=mode, side_effects=is_dest_op),
-            )
-
-            def load() -> ExpressionIndex:
-                # Use width 'w' (e.g. 1 for byte) for memory load/store element size
-                assert isinstance(op, Pointer)
-                return op.memory_helper()(w, ptr).lift(il)
-            def store(val: ExpressionIndex) -> None:
-                assert isinstance(op, Pointer)
-                op.memory_helper()(w, ptr).lift_assign(il, val)
-            def advance() -> None:
-                op_il_math = il.sub if reverse else il.add
-                # Advance pointer by element width 'w'
-                ptr.lift_assign(il, op_il_math(3, ptr.lift(il), il.const(3, w))) # ptr is 3 bytes
-        else: # Register operand
-            def load() -> ExpressionIndex:
-                return op.lift(il)
-            def store(val: ExpressionIndex) -> None:
-                op.lift_assign(il, val)
-            def advance() -> None: # No advancement for direct register operands in a loop
-                pass
-        return load, store, advance
-
-    w = op1.width()
-
-    load1, store1, adv1 = make_handlers(op1, True, dst_mode)
-    load2, store2, adv2 = make_handlers(op2, False, src_mode)
-
-    if clear_carry:
-        il.append(il.set_flag(CFlag, il.const(1, 0)))
-
-    overall_zero_acc_reg = TempReg(TempOverallZeroAcc, width=w)
-    overall_zero_acc_reg.lift_assign(il, il.const(w, 0))
-
-    # TempReg to store the result of the current byte's main arithmetic operation
-    byte_op_result_holder = TempReg(TempLoopByteResult, width=w)
-
-    with lift_loop(il): # loop_reg is 'I', controls number of iterations (bytes)
-        a = load1() # ExpressionIndex for current byte of op1
-        b = load2() # ExpressionIndex for current byte of op2
-
-        # This will hold the evaluated result of the current byte's operation
-        # before it's stored or used in overall_zero_acc.
-        current_byte_calculated_value_expr: ExpressionIndex
-
-        if bcd:
-            # BCD operations are complex; they read il.flag(CFlag) internally for incoming carry,
-            # perform BCD arithmetic, set CFlag and ZFlag (for the byte) based on BCD logic,
-            # and return an Operand (specifically a TempReg like TempBcdAddEmul or TempBcdSubEmul)
-            # which holds the BCD result of the current byte.
-            bcd_op_result_operand: Operand
-            if subtract: # DSBL
-                bcd_op_result_operand = bcd_sub_emul(il, w, a, b)
-            else: # DADL
-                bcd_op_result_operand = bcd_add_emul(il, w, a, b)
-
-            # The expression for the result of this byte's BCD operation
-            current_byte_calculated_value_expr = bcd_op_result_operand.lift(il)
-            # No need to assign to byte_op_result_holder if flags are fully set by bcd_emul
-            # and result is self-contained in its returned TempReg.
-            # The flags (C and Z for the byte) are set by set_flag calls within bcd_xxx_emul.
-        else: # Binary: ADCL, SBCL
-            # These operations use il.flag(CFlag) for incoming carry and set CZFlag for outgoing.
-            main_op_llil: ExpressionIndex
-
-            # Capture the incoming carry flag *before* this byte's main operation
-            initial_c_flag_expr = il.flag(CFlag)
-
-            if subtract: # SBCL: m = m - n - C_in. Implemented as m - (n + C_in)
-                         # The inner add (n + C_in) must NOT alter flags.
-                term_to_subtract = il.add(w, b, initial_c_flag_expr)
-                main_op_llil = il.sub(w, a, term_to_subtract, CZFlag) # This SUB sets C and Z flags
-            else: # ADCL: m = m + n + C_in. Implemented as m + (n + C_in)
-                  # The inner add (n + C_in) must NOT alter flags.
-                term_to_add = il.add(w, b, initial_c_flag_expr)
-                main_op_llil = il.add(w, a, term_to_add, CZFlag) # This ADD sets C and Z flags
-
-            # Execute the main operation and store its result in byte_op_result_holder.
-            # The flags (C and Z) are set when main_op_llil is evaluated as part of this set_reg.
-            byte_op_result_holder.lift_assign(il, main_op_llil)
-            current_byte_calculated_value_expr = byte_op_result_holder.lift(il) # = REG(TempLoopByteResult)
-
-        # Store the result for the current byte using the calculated value
-        store1(current_byte_calculated_value_expr)
-
-        # Accumulate for overall Zero flag check. This OR must not affect C/Z flags.
-        overall_zero_acc_reg.lift_assign(il, il.or_expr(w, overall_zero_acc_reg.lift(il), current_byte_calculated_value_expr))
-
-        adv1()
-        adv2()
-
-    # After loop, set the final Zero flag based on the accumulator
-    il.append(il.set_flag(ZFlag, il.compare_equal(w, overall_zero_acc_reg.lift(il), il.const(w, 0))))
-    # The Carry flag (FC) will hold the carry/borrow from the last byte's operation.
-
-
-class ADCL(ArithmeticInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst, src = self.operands()
-        # ADCL uses the incoming carry flag for the first byte.
-        lift_multi_byte(il, dst, src, clear_carry=False, pre=self._pre)
-
-class SBCL(ArithmeticInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst, src = self.operands()
-        # SBCL uses the incoming carry (borrow) flag for the first byte.
-        lift_multi_byte(il, dst, src, subtract=True, clear_carry=False, pre=self._pre)
-
-class DADL(ArithmeticInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst, src = self.operands()
-        # DADL does not use incoming carry for the first byte (implicitly 0).
-        lift_multi_byte(il, dst, src, clear_carry=True, bcd=True, reverse=True, pre=self._pre)
-
-class DSBL(ArithmeticInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst, src = self.operands()
-        # DSBL uses the incoming carry (borrow) flag for the first byte.
-        lift_multi_byte(
-            il,
-            dst,
-            src,
-            bcd=True,
-            subtract=True,
-            reverse=True,
-            clear_carry=False,
-            pre=self._pre,
-        )
-
-
-class LogicInstruction(Instruction): pass
-class AND(LogicInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.and_expr(1, il_arg1, il_arg2, ZFlag)
-class OR(LogicInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.or_expr(1, il_arg1, il_arg2, ZFlag)
-class XOR(LogicInstruction):
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.xor_expr(1, il_arg1, il_arg2, ZFlag)
-
-class CompareInstruction(Instruction): pass
-class TEST(CompareInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst_mode = get_addressing_mode(self._pre, 1) if self._pre else None
-        src_mode = get_addressing_mode(self._pre, 2) if self._pre else None
-        first, second = self.operands()
-        # FIXME: does it set the Z flag if any bit is set?
-        il.append(
-            il.set_flag(
-                ZFlag,
-                il.and_expr(3, first.lift(il, dst_mode), second.lift(il, src_mode)),
-            )
-        )
-
-class CMP(CompareInstruction):
-    def width(self) -> int:
-        return 1
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        dst_mode = get_addressing_mode(self._pre, 1) if self._pre else None
-        src_mode = get_addressing_mode(self._pre, 2) if self._pre else None
-        first, second = self.operands()
-        # FIXME: what's the proper width?
-        il.append(
-            il.sub(
-                self.width(),
-                first.lift(il, dst_mode),
-                second.lift(il, src_mode),
-                CZFlag,
-            )
-        )
-class CMPW(CMP):
-    def width(self) -> int:
-        return 2
-class CMPP(CMP):
-    def width(self) -> int:
-        return 3
-
-# FIXME: verify on real hardware, likely wrong
-class ShiftRotateInstruction(Instruction):
-    def shift_by(self, il: LowLevelILFunction) -> ExpressionIndex:
-        return il.const(1, 1)
-# bit rotation
-class ROR(ShiftRotateInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg1: ExpressionIndex) -> ExpressionIndex:
-        return il.rotate_right(1, il_arg1, self.shift_by(il), CZFlag)
-class ROL(ShiftRotateInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg1: ExpressionIndex) -> ExpressionIndex:
-        return il.rotate_left(1, il_arg1, self.shift_by(il), CZFlag)
-# bit shift
-class SHL(ShiftRotateInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg1: ExpressionIndex) -> ExpressionIndex:
-        return il.rotate_left_carry(1, il_arg1, self.shift_by(il),
-                                    il.flag(CFlag), CZFlag)
-class SHR(ShiftRotateInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg1: ExpressionIndex) -> ExpressionIndex:
-        return il.rotate_right_carry(1, il_arg1, self.shift_by(il),
-                                     il.flag(CFlag), CZFlag)
-
-# digit shift
-class DSLL(Instruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        imem_op, = self.operands()
-        # Ensure the operand is IMem8 as corrected in OPCODES
-        assert isinstance(imem_op, IMem8), f"DSLL operand should be IMem8, got {type(imem_op)}"
-
-        # current_addr_reg holds the internal memory address (e.g.,
-        # INTERNAL_MEMORY_START + n_offset)
-        # For DSLL, (n) is the MSB address.
-        current_addr_reg = TempReg(TempMultiByte1, width=3)  # Addresses are 3 bytes (20/24 bit)
-        mode = get_addressing_mode(self._pre, 1) if self._pre else None
-        current_addr_reg.lift_assign(
-            il, imem_op.lift_current_addr(il, pre=mode, side_effects=False)
-        )
-
-        # digit_carry_reg stores the high nibble of T (current byte) to be OR'd
-        # into the low nibble of S (shifted byte) in the next (less significant) iteration.
-        digit_carry_reg = TempReg(TempBcdDigitCarry, width=1)
-        # The first S is (T_low << 4) | 0. So initial carry is 0.
-        digit_carry_reg.lift_assign(il, il.const(1, 0))
-
-        overall_zero_acc_reg = TempReg(TempOverallZeroAcc, width=1)
-        overall_zero_acc_reg.lift_assign(il, il.const(1, 0))
-
-        mem_accessor = IMemHelper(width=1, value=current_addr_reg) # For load/store via current_addr_reg
-
-        with lift_loop(il): # loop_reg is 'I', decrements from initial value
-            current_byte_T = mem_accessor.lift(il)
-
-            # S = (T_low_nibble << 4) | digit_carry_from_PREVIOUS_T_high_nibble
-            T_low_nibble = il.and_expr(1, current_byte_T, il.const(1, 0x0F))
-            T_high_nibble = il.logical_shift_right(1, current_byte_T, il.const(1, 4))
-            # T_high_nibble = il.and_expr(1, T_high_nibble, il.const(1, 0x0F)) # Mask if needed
-
-            shifted_byte_S = il.or_expr(1,
-                                        il.shift_left(1, T_low_nibble, il.const(1, 4)),
-                                        digit_carry_reg.lift(il))
-
-            mem_accessor.lift_assign(il, shifted_byte_S)
-
-            # Update digit_carry_reg with T_high_nibble for the next iteration
-            digit_carry_reg.lift_assign(il, T_high_nibble)
-
-            overall_zero_acc_reg.lift_assign(il, il.or_expr(1, overall_zero_acc_reg.lift(il), shifted_byte_S))
-
-            # Decrement address pointer for DSLL (MSB to LSB)
-            current_addr_reg.lift_assign(il, il.sub(3, current_addr_reg.lift(il), il.const(3, 1)))
-
-        il.append(il.set_flag(ZFlag, il.compare_equal(1, overall_zero_acc_reg.lift(il), il.const(1, 0))))
-        # FC is not affected.
-
-class DSRL(Instruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        imem_op, = self.operands()
-        assert isinstance(imem_op, IMem8), f"DSRL operand should be IMem8, got {type(imem_op)}"
-
-        # For DSRL, (n) is the LSB address.
-        current_addr_reg = TempReg(TempMultiByte1, width=3)
-        mode = get_addressing_mode(self._pre, 1) if self._pre else None
-        current_addr_reg.lift_assign(
-            il, imem_op.lift_current_addr(il, pre=mode, side_effects=False)
-        )
-
-        # digit_carry_reg stores the low nibble of T (current byte) to be OR'd
-        # into the high nibble of S (shifted byte) in the next (more significant) iteration.
-        digit_carry_reg = TempReg(TempBcdDigitCarry, width=1)
-        # The first S is T >> 4. So initial carry is 0.
-        digit_carry_reg.lift_assign(il, il.const(1, 0))
-
-        overall_zero_acc_reg = TempReg(TempOverallZeroAcc, width=1)
-        overall_zero_acc_reg.lift_assign(il, il.const(1, 0))
-
-        mem_accessor = IMemHelper(width=1, value=current_addr_reg)
-
-        with lift_loop(il):
-            current_byte_T = mem_accessor.lift(il)
-
-            # S = T_high_nibble | (digit_carry_from_PREVIOUS_T_low_nibble << 4)
-            T_low_nibble = il.and_expr(1, current_byte_T, il.const(1, 0x0F))
-            T_high_nibble = il.logical_shift_right(1, current_byte_T, il.const(1, 4))
-            # T_high_nibble = il.and_expr(1, T_high_nibble, il.const(1, 0x0F))
-
-            shifted_byte_S = il.or_expr(1,
-                                        T_high_nibble,
-                                        il.shift_left(1, digit_carry_reg.lift(il), il.const(1, 4)))
-
-            mem_accessor.lift_assign(il, shifted_byte_S)
-
-            # Update digit_carry_reg with T_low_nibble for the next iteration
-            digit_carry_reg.lift_assign(il, T_low_nibble)
-
-            overall_zero_acc_reg.lift_assign(il, il.or_expr(1, overall_zero_acc_reg.lift(il), shifted_byte_S))
-
-            # Increment address pointer for DSRL (LSB to MSB)
-            current_addr_reg.lift_assign(il, il.add(3, current_addr_reg.lift(il), il.const(3, 1)))
-
-        il.append(il.set_flag(ZFlag, il.compare_equal(1, overall_zero_acc_reg.lift(il), il.const(1, 0))))
-        # FC is not affected.
-
-
-class IncDecInstruction(Instruction): pass
-class INC(IncDecInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg: ExpressionIndex) -> ExpressionIndex:
-        return il.add(1, il_arg, il.const(1, 1), ZFlag)
-class DEC(IncDecInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg: ExpressionIndex) -> ExpressionIndex:
-        return il.sub(1, il_arg, il.const(1, 1), ZFlag)
-
-class ExchangeInstruction(Instruction):
-    def lift_single_exchange(self, il: LowLevelILFunction, addr: int) -> None:
-        first, second = self.operands()
-        assert isinstance(first, HasWidth), f"Expected HasWidth, got {type(first)}"
-        width = first.width()
-        tmp = TempReg(TempExchange, width=width)
-        tmp.lift_assign(il, first.lift(il))
-        first.lift_assign(il, second.lift(il))
-        second.lift_assign(il, tmp.lift(il))
-
-    def encode(self, encoder: Encoder, addr: int) -> None:
-        op1, op2 = self.operands()
-        if isinstance(op1, IMemOperand) and isinstance(op2, IMemOperand):
-            pre_key = (op1.mode, op2.mode)
-            pre_byte = REVERSE_PRE_TABLE.get(pre_key)
-            if pre_byte is None:
-                raise ValueError(
-                    f"Invalid addressing mode combination for {self.name()}: {op1.mode.value} and {op2.mode.value}"
-                )
-            self._pre = pre_byte
-
-        super().encode(encoder, addr)
-class EX(ExchangeInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        self.lift_single_exchange(il, addr)
-# uses counter
-class EXL(ExchangeInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        with lift_loop(il):
-            self.lift_single_exchange(il, addr)
-
-class MiscInstruction(Instruction): pass
-class WAIT(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        with lift_loop(il):
-            # Wait is just an idle loop
-            pass
-
-class PMDF(MiscInstruction):
-    # FIXME: verify
-    def lift_operation2(self, il: LowLevelILFunction, il_arg1: ExpressionIndex, il_arg2: ExpressionIndex) -> ExpressionIndex:
-        return il.add(1, il_arg1, il_arg2)
-
-class SWAP(MiscInstruction):
-    def lift_operation1(self, il: LowLevelILFunction, il_arg1: ExpressionIndex) -> ExpressionIndex:
-        low = il.and_expr(1, il_arg1, il.const(1, 0x0F))
-        low = il.shift_left(1, low, il.const(1, 4))
-        high = il.and_expr(1, il_arg1, il.const(1, 0xF0))
-        high = il.logical_shift_right(1, high, il.const(1, 4))
-        return il.or_expr(1, low, high, ZFlag)
-
-class SC(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.set_flag(CFlag, il.const(1, 1)))
-class RC(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.set_flag(CFlag, il.const(1, 0)))
-
-# Timer Clear: sub-CG or main-CG timers are reset when STCL / MTCL of LCC are
-# set.
-# Divider ← D
-class TCL(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.intrinsic([], TCLIntrinsic, []))
-
-# System Clock Stop: halts main-CG of CPU
-# Execution can continue past HALT: ON, IRQ, KI pins
-# USR resets bits 0 to 2/5 to 0
-# SSR bit 2 and USR 3 and 4 are set to 1
-class HALT(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.intrinsic([], HALTIntrinsic, []))
-
-# System Clock Stop; Sub Clock Stop: main-CG and sub-CG of CPU are stopped
-# Execution can continue past OFF: ON, IRQ, KI pins
-class OFF(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.intrinsic([], OFFIntrinsic, []))
-
-# AKA `INT / Interrupt`
-# 1. Save context to system stack (S-stack), in this strict order:
-#      - PS  (Program Status)
-#      - PC  (Program Counter, high byte first, then low byte)
-#      - FLAG (Status Flags)
-#      - IMR (Interrupt Mask Register at FBH)
-#    (Total pushed = 5 bytes)
-#
-# 2. Load new PC and PS from fixed memory locations:
-#      - PC high-byte loaded from address FFFFBH
-#      - PC low-byte  loaded from address FFFFAH
-#      - PS loaded from address FFFFCH
-#
-# 3. After pushing IMR, bit 7 (IRM) of IMR is forcibly cleared to 0.
-class IR(MiscInstruction):
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        il.append(il.push(3, RegPC().lift(il)))
-        il.append(il.push(1, RegF().lift(il)))
-        imr, *rest = RegIMR().operands()
-        il.append(il.push(1, imr.lift(il)))
-        imr.lift_assign(il, il.and_expr(1, imr.lift(il), il.const(1, 0x7F)))
-
-        mem = EMemAddr(width=3)
-        mem.value = INTERRUPT_VECTOR_ADDR
-        il.append(il.jump(mem.lift(il)))
-
-# ACM bit 7, UCR + USR bits 0 to 2/5, IMR, SCR, SSR bit 2 are all reset to 0
-# USR bits 3 and 4 are set to 1
-class RESET(MiscInstruction):
-    def analyze(self, info: InstructionInfo, addr: int) -> None:
-        super().analyze(info, addr)
-        info.add_branch(BranchType.FunctionReturn)
-
-    def lift(self, il: LowLevelILFunction, addr: int) -> None:
-        mem = EMemAddr(width=3)
-        mem.value = ENTRY_POINT_ADDR
-        il.append(il.jump(mem.lift(il)))
-
-class UnknownInstruction(Instruction):
-    def name(self) -> str:
-        return f"??? ({self.opcode:02X})"
-
-
-OPCODES = {
-    # 00h
-    0x00: NOP,
-    0x01: RETI,
-    0x02: (JP_Abs, Opts(ops=[Imm16()])),
-    0x03: (JP_Abs, Opts(name="JPF", ops=[Imm20()])),
-    0x04: (CALL, Opts(ops=[Imm16()])),
-    0x05: (CALL, Opts(name="CALLF", ops=[Imm20()])),
-    0x06: RET,
-    0x07: RETF,
-    0x08: (MV, Opts(ops=[Reg("A"), Imm8()])),
-    0x09: (MV, Opts(ops=[Reg("IL"), Imm8()])),
-    0x0A: (MV, Opts(ops=[Reg("BA"), Imm16()])),
-    0x0B: (MV, Opts(ops=[Reg("I"), Imm16()])),
-    0x0C: (MV, Opts(ops=[Reg("X"), Imm20()])),
-    0x0D: (MV, Opts(ops=[Reg("Y"), Imm20()])),
-    0x0E: (MV, Opts(ops=[Reg("U"), Imm20()])),
-    0x0F: (MV, Opts(ops=[Reg("S"), Imm20()])),
-    # 10h
-    0x10: (JP_Abs, Opts(ops=[IMem20()])),
-    0x11: (JP_Abs, Opts(ops=[Reg3()])),
-    0x12: (JP_Rel, Opts(ops=[ImmOffset("+")])),
-    0x13: (JP_Rel, Opts(ops=[ImmOffset("-")])),
-    0x14: (JP_Abs, Opts(cond="Z", ops=[Imm16()])),
-    0x15: (JP_Abs, Opts(cond="NZ", ops=[Imm16()])),
-    0x16: (JP_Abs, Opts(cond="C", ops=[Imm16()])),
-    0x17: (JP_Abs, Opts(cond="NC", ops=[Imm16()])),
-    0x18: (JP_Rel, Opts(cond="Z", ops=[ImmOffset("+")])),
-    0x19: (JP_Rel, Opts(cond="Z", ops=[ImmOffset("-")])),
-    0x1A: (JP_Rel, Opts(cond="NZ", ops=[ImmOffset("+")])),
-    0x1B: (JP_Rel, Opts(cond="NZ", ops=[ImmOffset("-")])),
-    0x1C: (JP_Rel, Opts(cond="C", ops=[ImmOffset("+")])),
-    0x1D: (JP_Rel, Opts(cond="C", ops=[ImmOffset("-")])),
-    0x1E: (JP_Rel, Opts(cond="NC", ops=[ImmOffset("+")])),
-    0x1F: (JP_Rel, Opts(cond="NC", ops=[ImmOffset("-")])),
-    # 20h
-    0x20: UnknownInstruction,
-    0x21: PRE,
-    0x22: PRE,
-    0x23: PRE,
-    0x24: PRE,
-    0x25: PRE,
-    0x26: PRE,
-    0x27: PRE,
-    0x28: (PUSHU, Opts(ops=[Reg("A")])),
-    0x29: (PUSHU, Opts(ops=[Reg("IL")])),
-    0x2A: (PUSHU, Opts(ops=[Reg("BA")])),
-    0x2B: (PUSHU, Opts(ops=[Reg("I")])),
-    0x2C: (PUSHU, Opts(ops=[Reg("X")])),
-    0x2D: (PUSHU, Opts(ops=[Reg("Y")])),
-    0x2E: (PUSHU, Opts(ops=[RegF()])),
-    0x2F: (PUSHU, Opts(ops=[RegIMR()])),
-    # 30h
-    0x30: PRE,
-    0x31: PRE,
-    0x32: PRE,
-    0x33: PRE,
-    0x34: PRE,
-    0x35: PRE,
-    0x36: PRE,
-    0x37: PRE,
-    0x38: (POPU, Opts(ops=[Reg("A")])),
-    0x39: (POPU, Opts(ops=[Reg("IL")])),
-    0x3A: (POPU, Opts(ops=[Reg("BA")])),
-    0x3B: (POPU, Opts(ops=[Reg("I")])),
-    0x3C: (POPU, Opts(ops=[Reg("X")])),
-    0x3D: (POPU, Opts(ops=[Reg("Y")])),
-    0x3E: (POPU, Opts(ops=[RegF()])),
-    0x3F: (POPU, Opts(ops=[RegIMR()])),
-    # 40h
-    0x40: (ADD, Opts(ops=[Reg("A"), Imm8()])),
-    0x41: (ADD, Opts(ops=[IMem8(), Imm8()])),
-    0x42: (ADD, Opts(ops=[Reg("A"), IMem8()])),
-    0x43: (ADD, Opts(ops=[IMem8(), Reg("A")])),
-    0x44: (ADD, Opts(ops=[RegPair(size=2)])),
-    0x45: (ADD, Opts(ops=[RegPair(size=3)])),
-    0x46: (ADD, Opts(ops=[RegPair(size=1)])),
-    0x47: (PMDF, Opts(ops=[IMem8(), Imm8()])),
-    0x48: (SUB, Opts(ops=[Reg("A"), Imm8()])),
-    0x49: (SUB, Opts(ops=[IMem8(), Imm8()])),
-    0x4A: (SUB, Opts(ops=[Reg("A"), IMem8()])),
-    0x4B: (SUB, Opts(ops=[IMem8(), Reg("A")])),
-    0x4C: (SUB, Opts(ops=[RegPair(size=2)])),
-    0x4D: (SUB, Opts(ops=[RegPair(size=3)])),
-    0x4E: (SUB, Opts(ops=[RegPair(size=1)])),
-    0x4F: (PUSHS, Opts(ops=[RegF()])),
-    # 50h
-    0x50: (ADC, Opts(ops=[Reg("A"), Imm8()])),
-    0x51: (ADC, Opts(ops=[IMem8(), Imm8()])),
-    0x52: (ADC, Opts(ops=[Reg("A"), IMem8()])),
-    0x53: (ADC, Opts(ops=[IMem8(), Reg("A")])),
-    0x54: (ADCL, Opts(ops=[IMem8(), IMem8()])),
-    0x55: (ADCL, Opts(ops=[IMem8(), Reg("A")])),
-    0x56: (MVL, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM,
-                                        allowed_modes=[EMemRegMode.POSITIVE_OFFSET, EMemRegMode.NEGATIVE_OFFSET])])),
-    0x57: (PMDF, Opts(ops=[IMem8(), Reg("A")])),
-    0x58: (SBC, Opts(ops=[Reg("A"), Imm8()])),
-    0x59: (SBC, Opts(ops=[IMem8(), Imm8()])),
-    0x5A: (SBC, Opts(ops=[Reg("A"), IMem8()])),
-    0x5B: (SBC, Opts(ops=[IMem8(), Reg("A")])),
-    0x5C: (SBCL, Opts(ops=[IMem8(), IMem8()])),
-    0x5D: (SBCL, Opts(ops=[IMem8(), Reg("A")])),
-    0x5E: (MVL, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET,
-                                        allowed_modes=[EMemRegMode.POSITIVE_OFFSET, EMemRegMode.NEGATIVE_OFFSET])])),
-    0x5F: (POPS, Opts(ops=[RegF()])),
-    # 60h
-    0x60: (CMP, Opts(ops=[Reg("A"), Imm8()])),
-    0x61: (CMP, Opts(ops=[IMem8(), Imm8()])),
-    0x62: (CMP, Opts(ops=[EMemAddr(width=1), Imm8()])),
-    0x63: (CMP, Opts(ops=[IMem8(), Reg("A")])),
-    0x64: (TEST, Opts(ops=[Reg("A"), Imm8()])),
-    0x65: (TEST, Opts(ops=[IMem8(), Imm8()])),
-    0x66: (TEST, Opts(ops=[EMemAddr(width=1), Imm8()])),
-    0x67: (TEST, Opts(ops=[IMem8(), Reg("A")])),
-    0x68: (XOR, Opts(ops=[Reg("A"), Imm8()])),
-    0x69: (XOR, Opts(ops=[IMem8(), Imm8()])),
-    0x6A: (XOR, Opts(ops=[EMemAddr(width=1), Imm8()])),
-    0x6B: (XOR, Opts(ops=[IMem8(), Reg("A")])),
-    0x6C: (INC, Opts(ops=[Reg3()])),
-    0x6D: (INC, Opts(ops=[IMem8()])),
-    0x6E: (XOR, Opts(ops=[IMem8(), IMem8()])),
-    0x6F: (XOR, Opts(ops=[Reg("A"), IMem8()])),
-    # 70h
-    0x70: (AND, Opts(ops=[Reg("A"), Imm8()])),
-    0x71: (AND, Opts(ops=[IMem8(), Imm8()])),
-    0x72: (AND, Opts(ops=[EMemAddr(width=1), Imm8()])),
-    0x73: (AND, Opts(ops=[IMem8(), Reg("A")])),
-    0x74: (MV, Opts(ops=[Reg("A"), RegB()])),
-    0x75: (MV, Opts(ops=[RegB(), Reg("A")])),
-    0x76: (AND, Opts(ops=[IMem8(), IMem8()])),
-    0x77: (AND, Opts(ops=[Reg("A"), IMem8()])),
-    0x78: (OR, Opts(ops=[Reg("A"), Imm8()])),
-    0x79: (OR, Opts(ops=[IMem8(), Imm8()])),
-    0x7A: (OR, Opts(ops=[EMemAddr(width=1), Imm8()])),
-    0x7B: (OR, Opts(ops=[IMem8(), Reg("A")])),
-    0x7C: (DEC, Opts(ops=[Reg3()])),
-    0x7D: (DEC, Opts(ops=[IMem8()])),
-    0x7E: (OR, Opts(ops=[IMem8(), IMem8()])),
-    0x7F: (OR, Opts(ops=[Reg("A"), IMem8()])),
-    # 80h
-    0x80: (MV, Opts(ops=[Reg("A"), IMem8()])),
-    0x81: (MV, Opts(ops=[Reg("IL"), IMem8()])),
-    0x82: (MV, Opts(ops=[Reg("BA"), IMem16()])),
-    0x83: (MV, Opts(ops=[Reg("I"), IMem16()])),
-    0x84: (MV, Opts(ops=[Reg("X"), IMem20()])),
-    0x85: (MV, Opts(ops=[Reg("Y"), IMem20()])),
-    0x86: (MV, Opts(ops=[Reg("U"), IMem20()])),
-    0x87: (MV, Opts(ops=[Reg("S"), IMem20()])),
-    0x88: (MV, Opts(ops=[Reg("A"), EMemAddr(width=1)])),
-    0x89: (MV, Opts(ops=[Reg("IL"), EMemAddr(width=1)])),
-    0x8A: (MV, Opts(ops=[Reg("BA"), EMemAddr(width=2)])),
-    0x8B: (MV, Opts(ops=[Reg("I"), EMemAddr(width=2)])),
-    0x8C: (MV, Opts(ops=[Reg("X"), EMemAddr(width=3)])),
-    0x8D: (MV, Opts(ops=[Reg("Y"), EMemAddr(width=3)])),
-    0x8E: (MV, Opts(ops=[Reg("U"), EMemAddr(width=3)])),
-    0x8F: (MV, Opts(ops=[Reg("S"), EMemAddr(width=3)])),
-    # 90h
-    0x90: (MV, Opts(ops=[Reg("A"), EMemReg(width=1)])),
-    0x91: (MV, Opts(ops=[Reg("IL"), EMemReg(width=1)])),
-    0x92: (MV, Opts(ops=[Reg("BA"), EMemReg(width=2)])),
-    0x93: (MV, Opts(ops=[Reg("I"), EMemReg(width=2)])),
-    0x94: (MV, Opts(ops=[Reg("X"), EMemReg(width=3)])),
-    0x95: (MV, Opts(ops=[Reg("Y"), EMemReg(width=3)])),
-    0x96: (MV, Opts(ops=[Reg("U"), EMemReg(width=3)])),
-    0x97: SC,
-    0x98: (MV, Opts(ops=[Reg("A"), EMemIMem()])),
-    0x99: (MV, Opts(ops=[Reg("IL"), EMemIMem()])),
-    0x9A: (MV, Opts(ops=[Reg("BA"), EMemIMem()])),
-    0x9B: (MV, Opts(ops=[Reg("I"), EMemIMem()])),
-    0x9C: (MV, Opts(ops=[Reg("X"), EMemIMem()])),
-    0x9D: (MV, Opts(ops=[Reg("Y"), EMemIMem()])),
-    0x9E: (MV, Opts(ops=[Reg("U"), EMemIMem()])),
-    0x9F: RC,
-    # A0h
-    0xA0: (MV, Opts(ops=[IMem8(), Reg("A")])),
-    0xA1: (MV, Opts(ops=[IMem8(), Reg("IL")])),
-    0xA2: (MV, Opts(ops=[IMem16(), Reg("BA")])),
-    0xA3: (MV, Opts(ops=[IMem16(), Reg("I")])),
-    0xA4: (MV, Opts(ops=[IMem20(), Reg("X")])),
-    0xA5: (MV, Opts(ops=[IMem20(), Reg("Y")])),
-    0xA6: (MV, Opts(ops=[IMem20(), Reg("U")])),
-    0xA7: (MV, Opts(ops=[IMem20(), Reg("S")])),
-    0xA8: (MV, Opts(ops=[EMemAddr(width=1), Reg("A")])),
-    0xA9: (MV, Opts(ops=[EMemAddr(width=1), Reg("IL")])),
-    0xAA: (MV, Opts(ops=[EMemAddr(width=2), Reg("BA")])),
-    0xAB: (MV, Opts(ops=[EMemAddr(width=2), Reg("I")])),
-    0xAC: (MV, Opts(ops=[EMemAddr(width=3), Reg("X")])),
-    0xAD: (MV, Opts(ops=[EMemAddr(width=3), Reg("Y")])),
-    0xAE: (MV, Opts(ops=[EMemAddr(width=3), Reg("U")])),
-    0xAF: (MV, Opts(ops=[EMemAddr(width=3), Reg("S")])),
-    # B0h
-    0xB0: (MV, Opts(ops=[EMemReg(width=1), Reg("A")])),
-    0xB1: (MV, Opts(ops=[EMemReg(width=1), Reg("IL")])),
-    0xB2: (MV, Opts(ops=[EMemReg(width=2), Reg("BA")])),
-    0xB3: (MV, Opts(ops=[EMemReg(width=2), Reg("I")])),
-    0xB4: (MV, Opts(ops=[EMemReg(width=3), Reg("X")])),
-    0xB5: (MV, Opts(ops=[EMemReg(width=3), Reg("Y")])),
-    0xB6: (MV, Opts(ops=[EMemReg(width=3), Reg("U")])),
-    0xB7: (CMP, Opts(ops=[IMem8(), IMem8()])),
-    0xB8: (MV, Opts(ops=[EMemIMem(), Reg("A")])),
-    0xB9: (MV, Opts(ops=[EMemIMem(), Reg("IL")])),
-    0xBA: (MV, Opts(ops=[EMemIMem(), Reg("BA")])),
-    0xBB: (MV, Opts(ops=[EMemIMem(), Reg("I")])),
-    0xBC: (MV, Opts(ops=[EMemIMem(), Reg("X")])),
-    0xBD: (MV, Opts(ops=[EMemIMem(), Reg("Y")])),
-    0xBE: (MV, Opts(ops=[EMemIMem(), Reg("U")])),
-    0xBF: UnknownInstruction,
-    # C0h
-    0xC0: (EX, Opts(ops=[IMem8(), IMem8()])),
-    0xC1: (EX, Opts(name="EXW", ops=[IMem16(), IMem16()])),
-    0xC2: (EX, Opts(name="EXP", ops=[IMem20(), IMem20()])),
-    0xC3: (EXL, Opts(ops=[IMem8(), IMem8()])),
-    0xC4: (DADL, Opts(ops=[IMem8(), IMem8()])),
-    0xC5: (DADL, Opts(ops=[IMem8(), Reg("A")])),
-    0xC6: (CMPW, Opts(ops=[IMem16(), IMem16()])),
-    0xC7: (CMPP, Opts(ops=[IMem20(), IMem20()])),
-    0xC8: (MV, Opts(ops=[IMem8(), IMem8()])),
-    0xC9: (MV, Opts(name="MVW", ops=[IMem16(), IMem16()])),
-    0xCA: (MV, Opts(name="MVP", ops=[IMem20(), IMem20()])),
-    0xCB: (MVL, Opts(ops=[IMem8(), IMem8()])),
-    0xCC: (MV, Opts(ops=[IMem8(), Imm8()])),
-    0xCD: (MV, Opts(name="MVW", ops=[IMem16(), Imm16()])),
-    0xCE: TCL,
-    0xCF: (MVLD, Opts(ops=[IMem8(), IMem8()])),
-    # D0h
-    0xD0: (MV, Opts(ops=[IMem8(), EMemAddr(width=1)])),
-    0xD1: (MV, Opts(name="MVW", ops=[IMem16(), EMemAddr(width=2)])),
-    0xD2: (MV, Opts(name="MVP", ops=[IMem20(), EMemAddr(width=3)])),
-    0xD3: (MVL, Opts(ops=[IMem8(), EMemAddr(width=1)])),
-    0xD4: (DSBL, Opts(ops=[IMem8(), IMem8()])),
-    0xD5: (DSBL, Opts(ops=[IMem8(), Reg("A")])),
-    0xD6: (CMPW, Opts(ops_reversed=True, ops=[IMem16(), Reg3()])),
-    0xD7: (CMPP, Opts(ops_reversed=True, ops=[IMem20(), Reg3()])),
-    0xD8: (MV, Opts(ops=[EMemAddr(width=1), IMem8()])),
-    0xD9: (MV, Opts(name="MVW", ops=[EMemAddr(width=2), IMem16()])),
-    0xDA: (MV, Opts(name="MVP", ops=[EMemAddr(width=3), IMem20()])),
-    0xDB: (MVL, Opts(ops=[EMemAddr(width=1), IMem8()])),
-    0xDC: (MV, Opts(name="MVP", ops=[IMem20(), Imm20()])),
-    0xDD: (EX, Opts(ops=[Reg("A"), RegB()])),
-    0xDE: HALT,
-    0xDF: OFF,
-    # E0h
-    0xE0: (MV, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
-    0xE1: (MV, Opts(name="MVW",
-                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
-    0xE2: (MV, Opts(name="MVP",
-                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_IMEM)])),
-    # FIXME: verify width
-    0xE3: (MVL, Opts(ops_reversed=True, ops=[IMem8(),
-                                             EMemReg(width=1,
-                                                     allowed_modes=[EMemRegMode.POST_INC,
-                                                                    EMemRegMode.PRE_DEC])])),
-    0xE4: (ROR, Opts(ops=[Reg("A")])),
-    0xE5: (ROR, Opts(ops=[IMem8()])),
-    0xE6: (ROL, Opts(ops=[Reg("A")])),
-    0xE7: (ROL, Opts(ops=[IMem8()])),
-    0xE8: (MV, Opts(ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
-    0xE9: (MV, Opts(name="MVW",
-                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
-    0xEA: (MV, Opts(name="MVP",
-                    ops=[RegIMemOffset(order=RegIMemOffsetOrder.DEST_REG_OFFSET)])),
-    # FIXME: verify width
-    0xEB: (MVL, Opts(ops=[EMemReg(width=1), IMem8()])),
-    0xEC: (DSLL, Opts(ops=[IMem8()])),
-    0xED: (EX, Opts(ops=[RegPair(size=2)])),
-    0xEE: (SWAP, Opts(ops=[Reg("A")])),
-    0xEF: WAIT,
-    # F0h
-    0xF0: (MV, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
-    0xF1: (MV, Opts(name='MVW', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
-    0xF2: (MV, Opts(name='MVP', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
-    0xF3: (MVL, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_INT_MEM)])),
-    0xF4: (SHR, Opts(ops=[Reg("A")])),
-    0xF5: (SHR, Opts(ops=[IMem8()])),
-    0xF6: (SHL, Opts(ops=[Reg("A")])),
-    0xF7: (SHL, Opts(ops=[IMem8()])),
-    0xF8: (MV, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
-    0xF9: (MV, Opts(name='MVW', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
-    0xFA: (MV, Opts(name='MVP', ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
-    0xFB: (MVL, Opts(ops=[EMemIMemOffset(order=EMemIMemOffsetOrder.DEST_EXT_MEM)])),
-    0xFC: (DSRL, Opts(ops=[IMem8()])),
-    0xFD: (MV, Opts(ops=[RegPair(size=2)])),
-    0xFE: IR,
-    0xFF: RESET,
-}
